Move startPyroscopeContainer to utils function
