+++ pkg/types/main.go
@@ -5,11 +5,16 @@ import (
 	"fmt"
 	"os"
 	"regexp"
+	"runtime"
 	"strings"
+	"time"
 
 	"github.com/ansible/receptor/pkg/controlsvc"
 	"github.com/ansible/receptor/pkg/netceptor"
+	"github.com/ansible/receptor/pkg/utils"
 	"github.com/ansible/receptor/pkg/workceptor"
+	"github.com/grafana/pyroscope-go"
+	"gopkg.in/yaml.v2"
 )
 
 type NodeCfg struct {
@@ -86,4 +91,100 @@ func (cfg NodeCfg) Run() error {
 	workceptor.MainInstance.ListKnownUnitIDs() // Triggers a scan of unit dirs and restarts any that need it
 
 	return nil
+}
+
+type ReceptorPyroscopeCfg struct {
+	ApplicationName   string // e.g backend.purchases
+	Tags              map[string]string
+	ServerAddress     string // e.g http://pyroscope.services.internal:4040
+	BasicAuthUser     string // http basic auth user
+	BasicAuthPassword string // http basic auth password
+	TenantID          string // specify TenantId when using phlare multi-tenancy
+	UploadRate        string
+	ProfileTypes      []string
+	DisableGCRuns     bool // this will disable automatic runtime.GC runs between getting the heap profiles
+	HTTPHeaders       map[string]string
+}
+
+type UploadRate struct {
+	UploadRate time.Duration `yaml:"uploadRate"`
+}
+
+func (pyroscopeCfg ReceptorPyroscopeCfg) Init() error {
+	if pyroscopeCfg.ApplicationName == "" {
+		return nil
+	}
+
+	runtime.SetMutexProfileFraction(5)
+	runtime.SetBlockProfileRate(5)
+
+	_, err := pyroscope.Start(pyroscope.Config{
+		ApplicationName:   pyroscopeCfg.ApplicationName,
+		Tags:              pyroscopeCfg.Tags,
+		ServerAddress:     pyroscopeCfg.ServerAddress,
+		BasicAuthUser:     pyroscopeCfg.BasicAuthUser,
+		BasicAuthPassword: pyroscopeCfg.BasicAuthPassword,
+		TenantID:          pyroscopeCfg.TenantID,
+		UploadRate:        getUploadRate(pyroscopeCfg),
+		Logger:            pyroscope.StandardLogger,
+		ProfileTypes:      getProfileTypes(pyroscopeCfg),
+		DisableGCRuns:     pyroscopeCfg.DisableGCRuns,
+		HTTPHeaders:       pyroscopeCfg.HTTPHeaders,
+	})
+
+	if err != nil {
+		return err
+	} else {
+		hostIP := strings.Split(strings.Split(pyroscopeCfg.ServerAddress, ":")[1], "//")[1]
+		hostPort := strings.Split(pyroscopeCfg.ServerAddress, ":")[2]
+		utils.StartPyroscopeContainer(hostIP, hostPort)
+
+		return nil
+	}
+
+	// pyroscope.TagWrapper(context.Background(), pyroscope.Labels("test", "netceptor_main"), func(c context.Context) {
+	// 	netceptor.MainInstance = netceptor.New(context.Background(), cfg.ID)
+	// })
+}
+
+func getUploadRate(cfg ReceptorPyroscopeCfg) time.Duration {
+	if cfg.UploadRate == "" {
+		return 15 * time.Second
+	}
+	var uploadRate UploadRate
+	err := yaml.Unmarshal([]byte(cfg.UploadRate), &uploadRate)
+	if err != nil {
+		fmt.Println("failed to parse uploadRate from config file")
+	}
+
+	return uploadRate.UploadRate
+}
+
+func getProfileTypes(cfg ReceptorPyroscopeCfg) []pyroscope.ProfileType {
+	profileType := []pyroscope.ProfileType{
+		pyroscope.ProfileCPU,
+		pyroscope.ProfileAllocObjects,
+		pyroscope.ProfileAllocSpace,
+		pyroscope.ProfileInuseObjects,
+		pyroscope.ProfileInuseSpace,
+	}
+	if len(cfg.ProfileTypes) == 0 {
+		return profileType
+	}
+	for _, pt := range cfg.ProfileTypes {
+		switch pt {
+		case "ProfileGoroutines":
+			profileType = append(profileType, pyroscope.ProfileGoroutines)
+		case "ProfileMutexCount":
+			profileType = append(profileType, pyroscope.ProfileMutexCount)
+		case "ProfileMutexDuration":
+			profileType = append(profileType, pyroscope.ProfileMutexDuration)
+		case "ProfileBlockCount":
+			profileType = append(profileType, pyroscope.ProfileBlockCount)
+		case "ProfileBlockDuration":
+			profileType = append(profileType, pyroscope.ProfileBlockDuration)
+		}
+	}
+
+	return profileType
 }
